First steps of moving RF sweep out of manager.
